% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_indicator_list.R
\name{create_indicator_list}
\alias{create_indicator_list}
\title{Create a list of indicators based on the level1 field in the dhs_indicators function}
\usage{
create_indicator_list(level_1 = "all")
}
\arguments{
\item{level_1}{level1 indicators wanted.  If "all" then all indicators are returned.  The default is all}
}
\value{
a list of indicators based on level1
}
\description{
Create a list of indicators based on the level1 field in the dhs_indicators function
}
\examples{
 \dontrun{

 df <- create_indicator_list()}
}
