% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nat_disagg.R
\name{plot_nat_disagg}
\alias{plot_nat_disagg}
\title{Create bar graph of dhs indicator results}
\usage{
plot_nat_disagg(
  df,
  indicator,
  survey_year = 2022,
  breakdown = "Region",
  palette = "rocket",
  type = "Percent",
  reverse_ord = FALSE
)
}
\arguments{
\item{df}{input dataframe}

\item{indicator}{dhs indicador id to visualize}

\item{survey_year}{survey year for desired indicator results}

\item{breakdown}{indicator disaggregation type}

\item{palette}{applied color palette}

\item{type}{value type (percentage vs absolute)}

\item{reverse_ord}{sorting order for good results to bad}
}
\value{
a bar graph with dhs indicator results
}
\description{
Create bar graph of dhs indicator results
}
\examples{
 \dontrun{

 df <- plot_nat_disagg()}
}
