% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_batch.R
\name{process_batch}
\alias{process_batch}
\title{Function to create a dataset from DHS_indicator()}
\usage{
process_batch(start_idx, batch_size, country, breakdown, start, end)
}
\arguments{
\item{start_idx}{starting index of a group of indicators.  Grouped by a batch_size}

\item{batch_size}{number of indicators to be included in each batch}

\item{country}{countries - based on DHS_CountryCode}

\item{breakdown}{select: national, subnational, all, background}

\item{start}{start year of survey}

\item{end}{end year of survey}
}
\value{
a dataset with indicators
}
\description{
Function to create a dataset from DHS_indicator()
}
\examples{
 \dontrun{

 df <- process_batch()}
}
