% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_indicator_df.R
\name{pull_indicator_df}
\alias{pull_indicator_df}
\title{Create an RDS dataset for Mozambique based on a given list of indicators and years}
\usage{
pull_indicator_df(
  indicator,
  country = "MZ",
  breakdown_default = "all",
  start_year,
  end_year
)
}
\arguments{
\item{indicator}{list of indicators}

\item{country}{country}

\item{breakdown_default}{indicator disaggregation}

\item{start_year}{earliest year for surveys}

\item{end_year}{latest year for surveys}
}
\value{
a dataset of indicators and years
}
\description{
Create an RDS dataset for Mozambique based on a given list of indicators and years
}
\examples{
 \dontrun{

 df <- pull_indicator_df()}
}
